extensions: support multiarch bindtextdomain preload
This commit adds i386 support to the bindtextdomain.so preload on amd64
systems. On these systems, bindtextdomain.so is cross-compiled to i386
and the $LIB token is used in the LD_PRELOAD rpath so the linker automatically
preloads the correct library based on the arch of the binary it is linking.